// web/src/api/generated/model/budgetDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { BudgetLineInput } from './budgetLineInput';

export interface BudgetDto {
  fiscalYear: number;
  /** @nullable */
  costCenterId: string | null;
  start: string;
  end: string;
  lines: BudgetLineInput[];
}